# SQL Write Performance Improvements