@bheni mentioned that we did not flush concurrently while performing other writes, and this idea was taken and executed upon. Lowering the flush threshold from `65536` to `16384` also seemed to improve performance just a bit with the new parallelism, which was not true with the previous implementation.

For performance numbers, we have `Original`, which is how it performed before I did any of the write path merging stuff. We have `Thread Safe`, which are my most recent changes. Then we have `Concurrent`, which is this current PR. Just for fun, we have `No-op` last. Here I made all of the write functions on `TableEditor` immediately return, just to see the true overhead that `TableEditor` actually adds to any particular write path.

Importing a 10 million row `.psv`:
```
Original:    53.521s
Thread Safe: 107.816s
Concurrent:  76.974s
No-op:       33.110s
```
Here we can see that we've gained around a 40% improvement in performance over my last PR, which is fairly substantial. As a result, the `TableEditor` is only about 31% slower than the original implementation, which should be perfectly acceptable given the additional bookkeeping.

Since these changes are made to `TableEditor`, the benefits also work for the standard SQL path. Here are the results of 1 million `INSERT` statements:
```
Original:    67.074s
Thread Safe: 67.363s
Concurrent:  59.325s
No-op:       57.839s
```
We've improved performance by 13%! Not as large as the 40% before, but that's due to the additional overhead that the SQL engine brings. In fact, with how close `Concurrent` is to `No-op`, we're at the point that any further speed increases on the SQL path will have to come from the engine, since the `TableEditor` is performing quickly enough to keep up with the engine already.