@Hydrocharged 2020-06-01
Fun news! In the time since I uploaded this PR, I found another way to improve performance! In the `processFlush` function, we were calling `Table.GetRow`, which internally gets the row data. For that 10 million row `.psv`, that means it's getting that map 10 million times that it doesn't need to. On top of that, I make use of the row data myself, so I just moved that line up a bit, and reused that map for the row checks. I then re-ran the 10 million row import, and here's the updated results table:
```
Original:    53.521s
Thread Safe: 107.816s
Concurrent:  71.698s
No-op:       33.110s
```
We gained about 5 seconds! In percentage terms, we've gone from a 40% increase to a 50% increase over the thread safety rewrite, and we're only 34% slower compared to 44% slower (I said 31% last time because I couldn't math properly) from the original implementation. A pretty big increase for a small change!

Also, I've got no idea why these random `.proto` files were reformatted by the `format_repo.sh`.